requirements.txt
appdirs==1.4.4
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.0
Django==2.2.16
django-allauth==0.43.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
djangorestframework==3.12.1
djongo==1.3.3
dnspython==2.0.0
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
Markdown==3.3.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
pyparsing==2.4.7
pyrsistent==0.17.3
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.2.4
stripe==2.55.0
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.21
